Set the theme before doing i18n
During i18n printer resolution, if there is no locale cookie we render
the language selection screen.  When this screen is rendered it will
grab the users theme from the current context.  As a result of not
having added the theme to the context yet though, we get a panic.  To
mitigate this we set the theme before doing any internationalization.